add new classifier

+++ attachment_classifier/main.py
@@ -46,8 +46,7 @@ if __name__ == '__main__':
                 num_sanity_val_steps=1, 
                 auto_scale_batch_size=True, 
                 enable_model_summary=True,
-                logger=logger,
-                precision=16,)
+                logger=logger)
                 #callbacks=[EarlyStopping(monitor="val_loss")])
 
     if mode == 'train':
